# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FactoryMind API
  version: 1.0.0
  summary: Industrial IoT monitoring API for Italian manufacturing SMEs.
  description: |
    FactoryMind exposes REST endpoints for real-time production monitoring:
    facilities, lines, devices (machines), telemetry metrics, alerts, OEE
    (Overall Equipment Effectiveness), and user administration.

    Most endpoints require a Bearer access token. Error responses follow
    RFC 7807 Problem Details (`application/problem+json`). State-changing
    endpoints are rate-limited (60 req/min/IP default).
  contact:
    name: FactoryMind Engineering
    email: support@factorymind.example
    url: https://factorymind.example/support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3002
    description: Local development (docker-compose)
  - url: https://api.factorymind.example
    description: Production
...
paths:
...
  /api/users/login:
    post:
      tags: [auth]
      summary: Exchange email + password for access + refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12 }
      responses:
        '200':
          description: Token pair issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/users/token/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Problem' }
  /api/users/logout:
    post:
      tags: [auth]
      summary: Revoke all refresh tokens for the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Success }
  /api/users/me:
    get:
      tags: [auth]
      summary: Return the claims carried by the current access token
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
  /api/users:
    get:
      tags: [auth]
      summary: List users (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserList' }
    post:
      tags: [auth]
      summary: Create user (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserInput' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: RFC 7807 error response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    TokenPair:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        refresh_expires_at: { type: string, format: date-time }
        token_type: { type: string, enum: [Bearer] }
        user: { type: object }
    UserInput:
      type: object
      required: [email, full_name, password]
      properties:
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [admin, supervisor, operator, viewer] }
        facility_scope:
          type: array
          items: { type: string }
        password: { type: string, minLength: 12, maxLength: 128 }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        full_name: { type: string }
        role: { type: string }
        facility_scope:
          type: array
          items: { type: string }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
    UserList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/User' }
        count: { type: integer }